// medflow-erp/types.ts
export interface ImageFilters {
  brightness: number;
  contrast: number;
  exposure: number;
  sharpness: number;
  invert: boolean;
  sepia: boolean;
}